(/usr/share/games/fortunes/tao) % The beginning of the universe Is the mother of all things. Knowing the mother, on also knows the sons. Knowing the sons, yet remaining in touch with the mother, Brings freedom from the fear of death.
Keep your mouth shut,
Guard the senses,
And life is ever full.
Open your mouth,
Always be busy,
And life is beyond hope.

Seeing the small is insight;
Yielding to force is strength.
Using the outer light, return to insight,
And in this way be saved from harm.
This is learning constancy.