(/usr/share/games/fortunes/art) % The Great Movie Posters:
The Miracle of the Age!!!  A LION in your lap!  A LOVER in your arms!
		-- Bwana Devil (1952)

OVERWHELMING!  ELECTRIFYING!  BAFFLING!
Fire Can't Burn Them!  Bullets Can't Kill Them!  See the Unfolding of
the Mysteries of the Moon as Murderous Robot Monsters Descend Upon the
Earth!  You've Never Seen Anything Like It!  Neither Has the World!
	SEE... Robots from Space in All Their Glory!!!
		-- Robot Monster (1953)

1,965 pyramids, 5,337 dancing girls, one million swaying bullrushes,
802 scared bulls!
		-- The Egyptian (1954)